backend/requirements.txt
flask
flask-cors
python-dotenv
requests
PyGithub
google-generativeai
torch
transformers
werkzeug